refactor this branch and merge with main
